AUDIT-53: Shift the ReadAudit AOP logic block to service class to remove code dublicacy and make it standard design
