+++ api/src/main/java/org/openmrs/module/auditlogweb/api/ReadAuditService.java
@@ -13,24 +13,31 @@ import org.openmrs.annotation.Authorized;
 import org.openmrs.module.auditlogweb.ReadAuditLog;
 import org.openmrs.module.auditlogweb.api.utils.AuditLogConstants;
 
+import org.aspectj.lang.ProceedingJoinPoint;
+
 import java.util.Date;
 import java.util.List;
 
 public interface ReadAuditService {
-	
+
 	void logReadAudit(ReadAuditLog readAuditLog);
-	
+
 	void logReadAudits(List<ReadAuditLog> readAuditLogs);
-	
+
 	@Authorized(AuditLogConstants.VIEW_READ_AUDIT_LOGS)
 	List<ReadAuditLog> getReadAuditLogs(String eventType, String username, Date startDate, Date endDate, int page, int size);
-	
+
 	@Authorized(AuditLogConstants.VIEW_READ_AUDIT_LOGS)
 	long countReadAuditLogs(String eventType, String username, Date startDate, Date endDate);
-	
+
 	@Authorized(AuditLogConstants.VIEW_READ_AUDIT_LOGS)
 	ReadAuditLog getReadAuditLogById(Integer id);
-	
+
 	@Authorized(AuditLogConstants.VIEW_READ_AUDIT_LOGS)
 	List<ReadAuditLog> getRelatedReadLogs(String sessionId, int limit);
+
+	Object auditReadRequest(ProceedingJoinPoint joinPoint) throws Throwable;
+
+	void saveReadAuditRequest(String entityName, boolean isReadSuccess, Object result);
+
 }
